password.c up to PS1_GeneratePassword_nbContinue

+++ include/password.h
@@ -9,7 +9,7 @@ void FUN_801a17c8(u8 arg0);
 u8 PS1_GetLevelFromPassword(void);
 void PS1_GeneratePassword_LivesCount(u8 lives_count);
 u8 PS1_GetLivesFromPassword(void);
-void PS1_GeneratePassword_nbContinue(u32 param_1);
+void PS1_GeneratePassword_nbContinue(u8 param_1);
 u8 PS1_GetContinuesFromPassword(void);
 u8 PS1_ValidatePassword(void);
 u8 PS1_GeneratePassword(void);
